// src/app/core/services/model/area.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ItemPoint } from './itemPoint';
import { Room } from './room';

export interface Area { 
    rooms?: Array<Room>;
    items?: Array<ItemPoint>;
    name: string;
    pointAddressUuid?: string;
}