// src/app/core/services/model/itemPoint.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface ItemPoint { 
    uuid?: string;
    name?: string;
    type?: string;
    isDisabled?: boolean;
}